Refactor: Move `fuse_abi` into ll/
I think it makes sense that `ll` not depend on code from outside `ll`.  This
isn't quite true yet, but this is a step in the right direction.
