+++ src/reply.rs
@@ -7,14 +7,12 @@
 //! error() exactly once).
 
 #[cfg(target_os = "macos")]
-use crate::fuse_abi::fuse_getxtimes_out;
-use crate::fuse_abi::{fuse_attr, fuse_attr_out, fuse_entry_out, fuse_file_lock, fuse_kstatfs};
-use crate::fuse_abi::{
-    fuse_bmap_out, fuse_ioctl_out, fuse_lk_out, fuse_lseek_out, fuse_open_out, fuse_statfs_out,
-    fuse_write_out,
+use crate::ll::fuse_abi::fuse_getxtimes_out;
+use crate::ll::fuse_abi::{
+    fuse_attr, fuse_attr_out, fuse_bmap_out, fuse_create_out, fuse_dirent, fuse_direntplus,
+    fuse_entry_out, fuse_file_lock, fuse_getxattr_out, fuse_ioctl_out, fuse_kstatfs, fuse_lk_out,
+    fuse_lseek_out, fuse_open_out, fuse_out_header, fuse_statfs_out, fuse_write_out,
 };
-use crate::fuse_abi::{fuse_create_out, fuse_getxattr_out};
-use crate::fuse_abi::{fuse_dirent, fuse_direntplus, fuse_out_header};
 use libc::{c_int, EIO, S_IFBLK, S_IFCHR, S_IFDIR, S_IFIFO, S_IFLNK, S_IFREG, S_IFSOCK};
 use log::warn;
 use std::convert::{AsRef, TryInto};
